Fix a couple of bugs with `ahkpm install` (#78)

+++ src/core/installer.go
@@ -11,8 +11,8 @@ type Installer struct{}
 func (i Installer) Install(deps DependencyArray) {
 	pr := NewPackagesRepository()
 
-	lm := LockManifestFromCwd()
-	if deps.Equals(lm.Dependencies()) {
+	lm, err := LockManifestFromCwd()
+	if err == nil && deps.Equals(lm.Dependencies()) {
 		fmt.Println("No dependency changes found. Installing from lockfile.")
 		os.RemoveAll("ahkpm-modules")
 		for _, resolvedDep := range lm.Resolved {
